+++ docs/griptape-framework/structures/src/config_3.py
@@ -2,7 +2,8 @@ import os
 
 import boto3
 
-from griptape.config import AmazonBedrockDriverConfig, config
+from griptape.config import config
+from griptape.config.drivers import AmazonBedrockDriverConfig
 from griptape.structures import Agent
 
 config.drivers = AmazonBedrockDriverConfig(