+++ docs/griptape-framework/structures/src/config_5.py
@@ -1,4 +1,5 @@
-from griptape.config import AnthropicDriverConfig, config
+from griptape.config import config
+from griptape.config.drivers import AnthropicDriverConfig
 from griptape.structures import Agent
 
 config.drivers = AnthropicDriverConfig()