+++ docs/griptape-tools/official-tools/src/rest_api_tool_1.py
@@ -1,6 +1,7 @@
 from json import dumps
 
-from griptape.config import DriverConfig, config
+from griptape.config import config
+from griptape.config.drivers import DriverConfig
 from griptape.drivers import OpenAiChatPromptDriver
 from griptape.memory.structure import ConversationMemory
 from griptape.structures import Pipeline