+++ griptape/config/drivers/azure_openai_driver_config.py
@@ -4,7 +4,7 @@ from typing import Callable, Optional
 
 from attrs import Factory, define, field
 
-from griptape.config import DriverConfig
+from griptape.config.drivers import DriverConfig
 from griptape.drivers import (
     AzureOpenAiChatPromptDriver,
     AzureOpenAiEmbeddingDriver,